move sortTest board to game
