+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -20,6 +20,7 @@ class BoardTest {
                     blankRank + blankRank + blankRank + blankRank +
                     appendNewLine("pppppppp") +
                     appendNewLine("rnbqkbnr");
+
     @BeforeEach
     void setUp() {
         board = new Board();
@@ -55,23 +56,5 @@ class BoardTest {
         board.initBoard(boardString);
         assertEquals(board.calculateWhitePoint(), 38.0);
         assertEquals(board.calculateBlackPoint(), 38.0);
-    }
-
-    @Test
-    @DisplayName("기물의 점수를 기준으로 정렬할 수 있어야 한다.")
-    void sort() {
-        board.initBoard(boardString);
-        List<Piece> whitePieces = List.of(createWhiteQueen(), createWhiteRook(), createWhiteRook(), createWhiteBishop(), createWhiteBishop(),
-                createWhiteKnight(), createWhiteKnight(),
-                createWhitePawn(), createWhitePawn(), createWhitePawn(), createWhitePawn(),
-                createWhitePawn(), createWhitePawn(), createWhitePawn(), createWhitePawn(),
-                createWhiteKing());
-        List<Piece> blackPieces = List.of(createBlackQueen(), createBlackRook(), createBlackRook(), createBlackBishop(), createBlackBishop(),
-                createBlackKnight(), createBlackKnight(),
-                createBlackPawn(), createBlackPawn(), createBlackPawn(), createBlackPawn(),
-                createBlackPawn(), createBlackPawn(), createBlackPawn(), createBlackPawn(),
-                createBlackKing());
-        assertEquals(board.getSortedWhitePieces(false), whitePieces);
-        assertEquals(board.getSortedBlackPieces(false), blackPieces);
     }
 }
